Fixed server-side method failing.
For single responsibility the server-side method was being called to retrieve the latest therapeutic level record rather than re-implementing the logic on the client-side. However this led to a failure when trying to marshal the record's environment object giving the message "cannot marshal recursive dictionaries". It was time-consuming to debug the issue so it seemed pragmatic to take the hit to single responsibility instead.